% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFun_Function.R
\name{balance_samples}
\alias{balance_samples}
\title{Helper function to balance positive and negative samples in the training dataset.}
\usage{
balance_samples(df, pos_limit = 1000, neg_ratio = 4)
}
\arguments{
\item{df}{A data frame containing a binary response variable "Y" (0 or 1).}

\item{pos_limit}{Maximum number of positive samples allowed.}

\item{neg_ratio}{Ratio of negative to positive samples after downsampling.}
}
\value{
A balanced data frame with reduced sample size.
}
\description{
Helper function to balance positive and negative samples in the training dataset.
}
\keyword{internal}
