% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics_WAS_Obtain.R
\name{MultiOmics_WAS_Obtain}
\alias{MultiOmics_WAS_Obtain}
\title{Full MultiFun Framework Function}
\usage{
MultiOmics_WAS_Obtain(Phenotype, folder)
}
\arguments{
\item{Phenotype}{Character string specifying the trait of interest.}

\item{folder}{Path to the working directory where output will be saved.}
}
\value{
Returns path to processed GWAS file or creates necessary files in specified folder.
}
\description{
Input phenotype and run full MultiFun framework to obtain multi-omics summary statistics.
}
\examples{
\dontrun{
MultiOmics_WAS_Obtain(Phenotype = "AFib", folder = "/public/home/test_blank/tyd/Application/Result/AFib")
}
}
