% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFun_Function.R
\name{MultiOmics_Weight_MKL}
\alias{MultiOmics_Weight_MKL}
\title{This function integrates multiple omics summary datasets and computes weights}
\usage{
MultiOmics_Weight_MKL(Multi_Y_Data, X_Data)
}
\arguments{
\item{Multi_Y_Data}{A list of data frames, each representing one omics layer's summary statistics.}

\item{X_Data}{A data frame containing annotation features with at least column: POS.}
}
\value{
A numeric vector of weights corresponding to each omics layer.
}
\description{
This function integrates multiple omics summary datasets and computes weights
}
