% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFun_Function.R
\name{PriorCausalProbability}
\alias{PriorCausalProbability}
\title{This function trains a machine learning model using provided labels (\code{Y_Data}) and features (`X_Data),
and predicts the probability of each SNP being causal.}
\usage{
PriorCausalProbability(Y_Data, X_Data, Prediction_Model)
}
\arguments{
\item{Y_Data}{A data frame containing summary statistics with at least columns: POS, Index (Y).}

\item{X_Data}{A data frame containing annotation features with at least column: POS.}

\item{Prediction_Model}{A string indicating the prediction algorithm to use (e.g., "randomforest").}
}
\value{
A list containing:
\item{Probability}{Predicted probability for each SNP.}
\item{ROC}{ROC curve object from pROC::roc().}
}
\description{
This function trains a machine learning model using provided labels (\code{Y_Data}) and features (`X_Data),
and predicts the probability of each SNP being causal.
}
