% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuSiE_MultiFun_Function.R
\name{process_chr_prior}
\alias{process_chr_prior}
\title{Process All or Specific Regions on a Single Chromosome Using SuSiE with MultiFun Prior}
\usage{
process_chr_prior(
  Pheno,
  Prior_path,
  Prior_var,
  Summary_path,
  LDMatrix_path,
  Python_path,
  Chr_select,
  Region,
  Output_path
)
}
\arguments{
\item{Pheno}{A data frame containing phenotype information with columns "FID" and "Pheno".}

\item{Prior_path}{Base path to directory containing prior files (e.g., MultiFun probability files).}

\item{Prior_var}{Column name of the prior weight to use from prior files.}

\item{Summary_path}{Base path to directory containing GWAS summary statistic files.}

\item{LDMatrix_path}{Base path to directory containing LD matrix files.}

\item{Python_path}{Optional character string indicating the Python executable path needed to load sparse LD matrices.}

\item{Chr_select}{Character string specifying the chromosome number to process.}

\item{Region}{Character string specifying which region(s) to process; use "All" to run all regions.}

\item{Output_path}{Path to output directory where results will be saved.}
}
\value{
A data.table containing fine-mapping results for the processed chromosome.
Returns NULL if merging fails.
}
\description{
Processes all defined regions on a chromosome in parallel using multiple cores.
It wraps around \code{process_region} for per-region processing, supports both Linux and Windows,
and handles failed runs via retry mechanism. Results are merged, deduplicated by position (POS),
and saved to CSV.
}
\examples{
\dontrun{
process_chr_prior(
  Pheno = my_pheno_df,
  Prior_path = "/data/", #contain MultiFun folder
  Prior_var = "PriorCausalProbability_MKL",
  Summary_path = "/data/GWAS",
  LDMatrix_path = "data/UKBB_LD",
  Python_path = "/usr/bin/python3",
  Chr_select = "1",
  Region = "All",
  Output_path = "/results" #generate MultiFunSuSiE folder
)
}
}
