% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFun_Prior.R
\name{Process_MultiFun_chr}
\alias{Process_MultiFun_chr}
\title{Process MultiOmics data for a single chromosome to calculate prior probabilities.}
\usage{
Process_MultiFun_chr(All_chr_tmp, MultiOmics_summary, Annotation)
}
\arguments{
\item{All_chr_tmp}{Chromosome number (integer).}

\item{MultiOmics_summary}{A list of multi-omics summary statistics by chromosome.}

\item{Annotation}{A data frame containing annotation features with at least columns: CHR, POS, and other features.}
}
\value{
A list containing:
\item{MultiOmics_Weight}{Weights assigned to each omics layer.}
\item{MultiOmics_PriorProbability}{Data frame of predicted probabilities per omics and combined scores.}
}
\description{
This function processes one chromosome of multi-omics summary statistics and annotation data,
runs the MultiFun algorithm to compute prior causal SNP probabilities, and returns results.
}
