% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFun_Function.R
\name{RandomForest_ImportanceVar}
\alias{RandomForest_ImportanceVar}
\title{Extract important variables from Random Forest model.}
\usage{
RandomForest_ImportanceVar(Y, traindata)
}
\arguments{
\item{Y}{Response variable name (ignored but kept for interface consistency).}

\item{traindata}{Training data including the response variable "Y".}
}
\value{
Character vector of selected feature names.
}
\description{
Extract important variables from Random Forest model.
}
\keyword{internal}
