% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuSiE_MultiFun_Function.R
\name{SuSiE_Geno}
\alias{SuSiE_Geno}
\title{Run SuSiE with Genotype Data Using Prior Information}
\usage{
SuSiE_Geno(Geno, Pheno, Prior)
}
\arguments{
\item{Geno}{A matrix of genotype data (n x p), where n is sample size and p is number of SNPs.}

\item{Pheno}{A data frame containing phenotype data, must have column 'Pheno'.}

\item{Prior}{A vector or matrix of prior weights for each SNP.}
}
\value{
A data frame with two columns: 'SNP' and 'PIP' (Posterior Inclusion Probability).
}
\description{
Perform SuSiE fine-mapping using genotype data and prior weights.
}
\examples{
\dontrun{
SuSiE_Geno(Geno_bed, Pheno_data, Prior_weights)
}
}
